:muscle: Rename and merge `invoker.ts` and `base.ts` into `host.ts`
