+++ denops/@denops-private/host.ts
@@ -1,5 +1,38 @@
+import type { Disposable } from "https://deno.land/x/disposable@v1.2.0/mod.ts";
 import { is } from "https://deno.land/x/unknownutil@v3.11.0/mod.ts";
-import { Service } from "../service.ts";
+import { Service } from "./service.ts";
+
+/**
+ * Host (Vim/Neovim) which is visible from Service
+ */
+export interface Host extends Disposable {
+  /**
+   * Redraw text and cursor on Vim but Neovim.
+   */
+  redraw(force?: boolean): Promise<void>;
+
+  /**
+   * Call host function and return result
+   */
+  call(fn: string, ...args: unknown[]): Promise<unknown>;
+
+  /**
+   * Batch call host functions and return results and error
+   */
+  batch(
+    ...calls: (readonly [string, ...unknown[]])[]
+  ): Promise<readonly [unknown[], string]>;
+
+  /**
+   * Register invoker
+   */
+  register(invoker: Invoker): void;
+
+  /**
+   * Wait host close
+   */
+  waitClosed(): Promise<void>;
+}
 
 export class Invoker {
   #service: Service;
