+++ denops/@denops-private/service.ts
@@ -4,8 +4,7 @@ import type {
 } from "https://deno.land/x/denops_core@v5.0.0/mod.ts";
 import type { Disposable } from "https://deno.land/x/disposable@v1.2.0/mod.ts";
 import { toFileUrl } from "https://deno.land/std@0.210.0/path/mod.ts";
-import { Host } from "./host/base.ts";
-import { Invoker } from "./host/invoker.ts";
+import { Host, Invoker } from "./host.ts";
 import { DenopsImpl } from "./denops.ts";
 
 /**
